url update
Now the urls are created, checked, and saved to a file to used to download the data. However issues remain with performance and with sorting out the focus codes as not all are useful or show the radio burst. Will need to create a filter to help with this but for now good progress.